packages/rls/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zerotheft-rls"
version = "0.1.0"
description = "Tenant RLS context propagation and enforcement for zerotheft platform services"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "starlette>=0.37.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-dotenv>=0.5.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy-cockroachdb>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]
